seperating problems from solver
