+++ tests/lava/lib/optimization/solvers/bayesian/test_processes.py
@@ -5,11 +5,7 @@ import sys
 import unittest
 
 from lava.lib.optimization.solvers.bayesian.processes import (
-    BayesianOptimizer,
-    BaseObjectiveFunction,
-    SingleInputLinearFunction,
-    SingleInputNonLinearFunction,
-    DualContInputFunction,
+    BayesianOptimizer
 )
 
 class TestProcesses(unittest.TestCase):
@@ -216,161 +212,6 @@ class TestProcesses(unittest.TestCase):
                 # check the interval var containing the initialized boolean
                 self.assertEqual(opt.vars.num_iterations.shape, (1,))
                 self.assertEqual(opt.vars.initialized.get(), False)
-
-    def test_process_base_objective_func(self) -> None:
-        """test initialization of the BaseObjectiveFunction process"""
-        
-        num_params: int = 5
-        num_objectives: int = 2
-        
-        process = BaseObjectiveFunction(
-            num_params=num_params,
-            num_objectives=num_objectives
-        )
-
-        # checking shape of input and output ports
-        self.assertEqual(
-            process.x_in.shape,
-            (num_params, 1)
-        )
-        self.assertEqual(
-            process.y_out.shape,
-            (num_params + num_objectives, 1)
-        )
-
-        # checking shape of input variables
-        self.assertEqual(
-            process.vars.num_params.shape,
-            (1,)
-        )
-        self.assertEqual(
-            process.vars.num_objectives.shape,
-            (1,)
-        )
-
-        # checking values of input_variables
-        self.assertEqual(
-            process.vars.num_params.get(),
-            num_params
-        )
-        self.assertEqual(
-            process.vars.num_objectives.get(),
-            num_objectives
-        )
-
-    def test_process_single_input_linear_func(self) -> None:
-        """test initialization of the SingleInputLinearFunction process"""
-
-        num_params: int = 1
-        num_objectives: int = 1
-        
-        process = SingleInputLinearFunction()
-
-        # checking shape of input and output ports
-        self.assertEqual(
-            process.x_in.shape,
-            (num_params, 1)
-        )
-        self.assertEqual(
-            process.y_out.shape,
-            (num_params + num_objectives, 1)
-        )
-
-        # checking shape of input variables
-        self.assertEqual(
-            process.vars.num_params.shape,
-            (1,)
-        )
-        self.assertEqual(
-            process.vars.num_objectives.shape,
-            (1,)
-        )
-
-        # checking values of input_variables
-        self.assertEqual(
-            process.vars.num_params.get(),
-            num_params
-        )
-        self.assertEqual(
-            process.vars.num_objectives.get(),
-            num_objectives
-        )
-
-    def test_process_single_input_nonlinear_func(self) -> None:
-        """test initialization of the SingleInputNonLinearFunction process"""
-
-        num_params: int = 1
-        num_objectives: int = 1
-        
-        process = SingleInputNonLinearFunction()
-
-        # checking shape of input and output ports
-        self.assertEqual(
-            process.x_in.shape,
-            (num_params, 1)
-        )
-        self.assertEqual(
-            process.y_out.shape,
-            (num_params + num_objectives, 1)
-        )
-
-        # checking shape of input variables
-        self.assertEqual(
-            process.vars.num_params.shape,
-            (1,)
-        )
-        self.assertEqual(
-            process.vars.num_objectives.shape,
-            (1,)
-        )
-
-        # checking values of input_variables
-        self.assertEqual(
-            process.vars.num_params.get(),
-            num_params
-        )
-        self.assertEqual(
-            process.vars.num_objectives.get(),
-            num_objectives
-        )
-
-    def test_process_dual_cont_input_func(self) -> None:
-        """test initialization of the DualContInputFunction process"""
-
-        num_params: int = 2
-        num_objectives: int = 1
-        
-        process = DualContInputFunction()
-
-        # checking shape of input and output ports
-        self.assertEqual(
-            process.x_in.shape,
-            (num_params, 1)
-        )
-        self.assertEqual(
-            process.y_out.shape,
-            (num_params + num_objectives, 1)
-        )
-
-        # checking shape of input variables
-        self.assertEqual(
-            process.vars.num_params.shape,
-            (1,)
-        )
-        self.assertEqual(
-            process.vars.num_objectives.shape,
-            (1,)
-        )
-
-        # checking values of input_variables
-        self.assertEqual(
-            process.vars.num_params.get(),
-            num_params
-        )
-        self.assertEqual(
-            process.vars.num_objectives.get(),
-            num_objectives
-        )
 
 
 if __name__ == "__main__":
